BankersAlgorithm
Implemented resource allocation and demonstrated deadlock avoidance with the Banker's algorithm. The program uses a claim graph consisting of processes, multi-unit resources, request edges, allocation edges, and claim edges to represent the state of allocated resources to processes.

The graph can be represented by a set of arrays:

Resource array: an m-element 1-dimensional array, where m is the number of resources, and each entry resource[j] records the total number of units of resource j.
Available array: an m-element 1-dimensional array, where m is the number of resources and each entry available[j] records the number of units of resource j that are available.
Max claims array: an nxm-element 2-dimensional array, where m is the number of resources and n is the number of processes, and each entry maxclaim[i][j] contains an integer that records the maximum number of units of resource j that process i may ever request.
Allocation array: an nxm-element 2-dimensional array, where m is the number of resources and n is the number of processes, and each entry allocated[i][j] contains an integer that records the number of units of resource j that process I has actually been allocated.
Need array: an nxm 2-dimensional array, where m is the number of resources and n is the number of processes, and each entry need[i][j] contains an integer that records the number of units of resource j that process I may need in the future.